backend/requirements.txt
Django
Django==4.1
djangorestframework==3.13.1
requests==2.28.1
psycopg2-binary==2.9.3
django-cors-headers==3.13.0